requirements.txt
httpx
beautifulsoup4
python-dotenv
google-generativeai
notion-client